Moved operator criteria to separate interfaces. Remove dataview method from trigger, it's not accessible in this form.
